% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gradients.R
\name{create_smooth_gradient}
\alias{create_smooth_gradient}
\title{Create a gradient}
\usage{
create_smooth_gradient(R, G, B, incr = 255)
}
\arguments{
\item{R}{A single parameter function for calculating the red value.}

\item{G}{A single parameter function for calculating the green value.}

\item{B}{A single parameter function for calculating the blue value.}

\item{incr}{The number of steps to take in the gradient. Defaults to 255.}
}
\description{
This function creates an RGB gradient based on a series of functions
}
\examples{
R <- function(x) {
 0.237 - 2.13 * x + 26.92 * x^2 - 65.5*x^3 + 63.5*x^4 - 22.36*x^5
}
G <- function(x) {
 ((0.572 + 1.524 * x - 1.811 * x^2)/(1 - 0.291 * x + 0.1574 * x^2))^2
}
B <- function(x) {
 1/(1.579 - 4.03*x + 12.92*x^2 - 31.4*x^3 + 48.6*x^4 - 23.36*x^5)
}
create_smooth_gradient(R, G, B)
}
\keyword{colors}
\keyword{palette}

