% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gradients.R
\name{gradient.black}
\alias{gradient.black}
\title{Create a black gradient}
\usage{
gradient.black(incr = 255)
}
\arguments{
\item{incr}{The number of steps to take in the gradient. Defaults to 255.}
}
\description{
This function creates a smooth, color blind-friendly, black-white gradient
}
\examples{
gradient.smooth.black(5)
}
\keyword{colors}
\keyword{palette}

