% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gradients.R
\name{gradient.blue}
\alias{gradient.blue}
\title{Create a mono blue gradient}
\usage{
gradient.blue(incr = 255)
}
\arguments{
\item{incr}{The number of steps to take in the gradient. Defaults to 255.}
}
\description{
This function creates a smooth, color blind-friendly, blue-white gradient
}
\examples{
gradient.smooth.blue(R, G, B)
}
\keyword{colors}
\keyword{palette}

