% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gradients.R
\name{gradient.brown2blue}
\alias{gradient.brown2blue}
\title{Create a brown-blue gradient}
\usage{
gradient.brown2blue()
}
\description{
This function returns a color blind-friendly, brown-blue gradient
}
\examples{
gradient.brown2blue()
}
\keyword{colors}
\keyword{palette}

