% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gradients.R
\name{gradient.focus}
\alias{gradient.focus}
\title{Create a black-white gradient with a highlight}
\usage{
gradient.focus(highlight = "red")
}
\arguments{
\item{highlight}{The color to use as a highlight.
Default is "red". Options are "green" or "blue".}
}
\description{
This function returns a black to white gradient with a highlight
}
\examples{
gradient.focus(highlight = "red")
}
\keyword{colors}
\keyword{palette}

