% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gradients.R
\name{gradient.gray2blue}
\alias{gradient.gray2blue}
\title{Create a gray-blue gradient}
\usage{
gradient.gray2blue()
}
\description{
This function returns a color blind-friendly, gray-blue gradient
}
\examples{
gradient.gray2blue()
}
\keyword{colors}
\keyword{palette}

