% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gradients.R
\name{gradient.green}
\alias{gradient.green}
\title{Create a mono green gradient}
\usage{
gradient.green(incr = 255)
}
\arguments{
\item{incr}{The number of steps to take in the gradient. Defaults to 255.}
}
\description{
This function creates a smooth, color blind-friendly, green-white gradient
}
\examples{
gradient.smooth.green(R, G, B)
}
\keyword{colors}
\keyword{palette}

