% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gradients.R
\name{gradient.green2purple}
\alias{gradient.green2purple}
\title{#' Create a green-purple gradient}
\usage{
gradient.green2purple()
}
\description{
This function returns a color blind-friendly, green-purple gradient
}
\examples{
gradient.green2purple()
}
\keyword{colors}
\keyword{palette}

