% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gradients.R
\name{gradient.purple}
\alias{gradient.purple}
\title{Create a mono purple gradient}
\usage{
gradient.purple(incr = 255)
}
\arguments{
\item{incr}{The number of steps to take in the gradient. Defaults to 255.}
}
\description{
This function creates a smooth, color blind-friendly, purple-white gradient
}
\examples{
gradient.smooth.purple(R, G, B)
}
\keyword{colors}
\keyword{palette}

