% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gradients.R
\name{gradient.rainbow}
\alias{gradient.rainbow}
\title{Create a rainbow gradient}
\usage{
gradient.rainbow(incr = 255)
}
\arguments{
\item{incr}{The number of steps to take in the gradient. Defaults to 255.}
}
\description{
This function creates a smooth, color blind-friendly, rainbow gradient
}
\examples{
gradient.smooth.red(R, G, B)
}
\keyword{colors}
\keyword{palette}

