% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gradients.R
\name{gradient.red}
\alias{gradient.red}
\title{Create a mono red gradient}
\usage{
gradient.red(incr = 255)
}
\arguments{
\item{incr}{The number of steps to take in the gradient. Defaults to 255.}
}
\description{
This function creates a smooth, color blind-friendly, red-white gradient
}
\examples{
gradient.smooth.red(255)
}
\keyword{colors}
\keyword{palette}

