% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gradients.R
\name{gradient.red2blue}
\alias{gradient.red2blue}
\title{Create a red-blue gradient}
\usage{
gradient.red2blue(incr = 255)
}
\arguments{
\item{incr}{The number of steps to take in the gradient. Defaults to 255.}
}
\description{
This function creates a smooth, color blind-friendly, red-blue gradient
}
\examples{
gradient.smooth.red2Blue(255)
}
\keyword{colors}
\keyword{palette}

