% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Palettes.R
\name{palette.dark}
\alias{palette.dark}
\title{Color palette from gplots rich.colors}
\usage{
palette.dark(n.colors = 12)
}
\arguments{
\item{n.colors}{The number of colors in the palette. Default=12. Options={6,8,10}}
}
\description{
This function creates a color blind-friendly, rainbow palette
}
\examples{
palette.dark()
}
\keyword{colors}
\keyword{palette}

