% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Palettes.R
\name{palette.qual}
\alias{palette.qual}
\title{Qualitative color palette by Paul Tol}
\usage{
palette.qual(n.colors = 12)
}
\arguments{
\item{n.colors}{The number of colors in the palette. Default = 12. Options = {1-12}}
}
\description{
This function creates a color blind-friendly, rainbow palette
}
\examples{
palette.qual(10)
}
\keyword{colors}
\keyword{palette}

