% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Palettes.R
\name{palette.rainbow.alt}
\alias{palette.rainbow.alt}
\title{Another color-blind friendly qualitative scheme}
\usage{
palette.rainbow.alt()
}
\description{
This function creates a color blind-friendly, rainbow palette
}
\examples{
palette.qual(10)
}
\keyword{color-blind}
\keyword{colors}
\keyword{palette}
\keyword{rainbow}

