% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Palettes.R
\name{palette.rainbow}
\alias{palette.rainbow}
\title{Create a rainbow palette}
\usage{
palette.rainbow(n.colors = 12)
}
\arguments{
\item{n.colors}{The number of colors in the palette. Default = 12. Options = {6, 8, 10, 12, 14, 15, 18, 21}}
}
\description{
This function creates a color blind-friendly, rainbow palette
}
\examples{
palette.rainbow(21)
}
\keyword{colors}
\keyword{palette}

