% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Palettes.R
\name{palette.rich}
\alias{palette.rich}
\title{Color palette from gplots rich.colors}
\usage{
palette.rich(n.colors = 12)
}
\arguments{
\item{n.colors}{The number of colors in the palette. Default=12. Options={6,8,10,12}}
}
\description{
This function creates a color blind-friendly palette
}
\examples{
palette.rich(10)
}
\keyword{colors}
\keyword{palette}

