% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShowPalette.R
\name{plot_palettes}
\alias{plot_palettes}
\title{Show palette with colorblind simulations}
\usage{
plot_palettes(col_list, border = NA, ...)
}
\arguments{
\item{col_list}{A list of color vectors}

\item{border}{The color of the border separating each color. Defaults to NA.}
}
\description{
This function plots multiple vector of colors side by side
}
\examples{
a <- list(gradient.green(8), gradient.blue(8), gradient.red(8))
plot_palettes(a, border = "light gray")
}
\keyword{colors}
\keyword{palette}

