% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShowPalette.R
\name{remcolor.showPaletteColorblind}
\alias{remcolor.showPaletteColorblind}
\title{Show palette with colorblind simulations}
\usage{
remcolor.showPaletteColorblind(col, border = NA, ...)
}
\arguments{
\item{col}{The color vector}

\item{border}{The color of the border separating each color. Defaults to NA.}
}
\description{
This function plots vector of colors side by side
}
\examples{
showPalette(c("#005000","#008600","#00BB00"), border="light gray")
}
\keyword{colors}
\keyword{palette}

