% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShowPalette.R
\name{remcolor.showPalettes}
\alias{remcolor.showPalettes}
\title{Show palettes}
\usage{
remcolor.showPalettes(col_list, border = NA, show.rgb = T, ...)
}
\arguments{
\item{border}{The color of the border separating each color. Defaults to NA.}

\item{col}{The color vector}
}
\description{
This function plots vector of colors side by side
}
\examples{
showPalette(c("#005000","#008600","#00BB00"), border="light gray")
}
\keyword{colors}
\keyword{palette}

