% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateColorBlind.R
\name{remcolor.simulate_colorblind_blue}
\alias{remcolor.simulate_colorblind_blue}
\title{Tritanope (blue-blind) version of palette (0.003% population)}
\usage{
remcolor.simulate_colorblind_blue(col)
}
\arguments{
\item{col}{The color vector}
}
\description{
Generate a set of hex colors that simulate blue color blindness
}
\examples{
remcolor.simulate_colorblind_blue(c("#005000","#008600","#00BB00"), border="light gray")
}
\keyword{colorblind}
\keyword{colors,}
\keyword{palette,}

