% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateColorBlind.R
\name{remcolor.simulate_colorblind_hex}
\alias{remcolor.simulate_colorblind_hex}
\title{Simulate colorblindness for a color}
\usage{
remcolor.simulate_colorblind_hex(hex_str, cvd)
}
\arguments{
\item{hex_str}{The hex color string, e.g., '#FF0000'}

\item{cvd}{A vector (3x3--9 values) specifying the color vision deficiency transform matrix}
}
\description{
Generate a hex color simulating colorblindness.
Algorithm from http://www.daltonize.org/
}
\examples{
remcolor.simulate_colorblind_green(c("#005000","#008600","#00BB00"), border="light gray")
}
\keyword{colorblind}
\keyword{colors,}
\keyword{palette,}

