% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateColorBlind.R
\name{remcolor.simulate_colorblind_red}
\alias{remcolor.simulate_colorblind_red}
\title{Protanope (red-blind) version of palette (1% men)}
\usage{
remcolor.simulate_colorblind_red(col)
}
\arguments{
\item{col}{The color vector}
}
\description{
Generate a set of hex colors that simulate green color blindness
}
\examples{
remcolor.simulate_colorblind_red(c("#005000","#008600","#00BB00"), border="light gray")
}
\keyword{colorblind}
\keyword{colors,}
\keyword{palette,}

