% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShowPalette.R
\name{show_palette}
\alias{show_palette}
\title{Show palette}
\usage{
show_palette(col, border = NA, ...)
}
\arguments{
\item{col}{The color vector}

\item{border}{The color of the border separating each color. Defaults to NA.}
}
\description{
This function plots vector of colors side by side
}
\examples{
show_palette(c("#005000", "#008600", "#00BB00"), border = "light gray")
}
\keyword{colors}
\keyword{palette}

